Main discord bot code.
This code extracts the Spotify playlist data and passes it to the discord bot, who prints it in the server.
Sadly, Rythm bot doesn't listen to other bots. So we have to make it with another perspective.
It doesn't get multiple commands in a single message either.
The solution would be generating a youtube playlist and create the command so the user can forward it to the chat and make Rythm bot playit.